feat(reportes): implementar MOD-NF-03 - CU-87, CU-88 y CU-89 completos

+++ src/main/java/com/app/idoneos/service/Pago/PagoService.java
@@ -111,7 +111,7 @@ public class PagoService {
 
         String paymentId = llamarMercadoPagoAPI(monto, emailPagador, nombrePagador);
 
-        Pago pago = new Pago(monto, inscripcion, acreditado);
+        Pago pago = new Pago(monto.floatValue(), inscripcion, acreditado);
         pago.setEmailPagador(emailPagador);
         pago.setNombrePagador(nombrePagador);
         pago.setMetodoPago(metodo);
